requirements.txt

binance-futures-connector==4.0.0
ntplib==0.4.0
numpy==1.26.4
pandas==2.2.2
pybit==5.7.0
python-binance==1.0.19
python-telegram-bot==21.1.1
tabulate==0.9.0
